修改bug

+++ tool-pdf/src/main/java/com/allen/tool/pdf/PDFTool.java
@@ -145,7 +145,7 @@ public class PDFTool {
 
                 str = str.replace(",", "");
 
-                int sInt = Integer.parseInt(str);
+                long sInt = Long.parseLong(str);
                 if ( sInt != 0 ) {
                     continue;
                 }
